# openapi/unified-openapi.yaml
openapi: 3.1.0
info:
  title: UserManagement
  version: 1.0.0
...
paths:
...
  /api/v1/meetings/{id}/recordings:stop:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    post:
      tags:
        - Recordings
      summary: Stop recording a meeting
      operationId: stopRecording
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MeetingManagement_JsendResponseVoid'
  /api/v1/meetings/{id}/recordings:start:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    post:
      tags:
        - Recordings
      summary: Start recording a meeting
      operationId: startRecording
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/MeetingManagement_JsendResponseRecordingResponse
...
  /api/v1/users/{userId}/meetings:filter:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    get:
      tags:
        - User Meetings
      summary: List participated meetings with optional status filter
      operationId: listParticipatedMeetings
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: pageToken
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/MeetingManagement_JsendResponseCursorScrollResponseMeetingResponse
  /api/v1/users/{userId}/meetings/{meetingId}:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    get:
      tags:
        - User Meetings
      summary: Get participated meeting detail
      operationId: getParticipatedMeetingDetail
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/MeetingManagement_JsendResponseMeetingDetailResponse
  /api/v1/recordings/{id}:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    get:
      tags:
        - Recordings
      summary: Get recording by ID
      operationId: getRecording
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/MeetingManagement_JsendResponseRecordingResponse
...
  /api/v1/meetings/{id}/recordings:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    get:
      tags:
        - Recordings
      summary: List recordings for a meeting
      operationId: listMeetingRecordings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: pageToken
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/MeetingManagement_JsendResponseCursorScrollResponseRecordingResponse
...
components:
  schemas:
...
    MeetingManagement_MeetingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hostId:
          type: string
          format: uuid
        shortCode:
          type: string
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        type:
          type: string
          enum:
            - INSTANT
            - SCHEDULED
        status:
          type: string
          enum:
            - SCHEDULED
            - LIVE
            - ENDED
            - CANCELLED
        settings:
          $ref: '#/components/schemas/MeetingManagement_MeetingSettingsResponse'
        createdAt:
          type: string
          format: date-time
    MeetingManagement_MeetingSettingsResponse:
      type: object
      properties:
        admissionPolicy:
          type: string
        joinRequestTimeoutSeconds:
          type: integer
          format: int32
        allowGuest:
          type: boolean
        muteOnEntry:
          type: boolean
        maxParticipants:
          type: integer
          format: int32
        recordingEnabled:
          type: boolean
        screenShareMode:
          type: string
        chatEnabled:
          type: boolean
        requirePassword:
          type: boolean
...
    MeetingManagement_JsendResponseVoid:
      type: object
      properties:
        status:
          type: string
        data: {}
        message:
          type: string
...
    MeetingManagement_JsendResponseRecordingResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: '#/components/schemas/MeetingManagement_RecordingResponse'
        message:
          type: string
    MeetingManagement_RecordingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        meetingId:
          type: string
          format: uuid
        fileUrl:
          type: string
        thumbnailUrl:
          type: string
        status:
          type: string
          enum:
            - PENDING
            - RECORDING
            - COMPLETED
            - FAILED
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        durationSeconds:
          type: integer
          format: int32
        fileSizeBytes:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
...
    MeetingManagement_CursorScrollResponseMeetingResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MeetingManagement_MeetingResponse'
        size:
          type: integer
          format: int32
        nextPageToken:
          type: string
    MeetingManagement_JsendResponseCursorScrollResponseMeetingResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: >-
            #/components/schemas/MeetingManagement_CursorScrollResponseMeetingResponse
        message:
          type: string
    MeetingManagement_InviteeResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        status:
          type: string
          enum:
            - PENDING
            - ACCEPTED
            - DECLINED
        invitedAt:
          type: string
          format: date-time
        respondedAt:
          type: string
          format: date-time
    MeetingManagement_JsendResponseMeetingDetailResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: '#/components/schemas/MeetingManagement_MeetingDetailResponse'
        message:
          type: string
    MeetingManagement_MeetingDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hostId:
          type: string
          format: uuid
        shortCode:
          type: string
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        type:
          type: string
          enum:
            - INSTANT
            - SCHEDULED
        status:
          type: string
          enum:
            - SCHEDULED
            - LIVE
            - ENDED
            - CANCELLED
        settings:
          $ref: '#/components/schemas/MeetingManagement_MeetingSettingsResponse'
        createdAt:
          type: string
          format: date-time
        participants:
          type: array
          items:
            $ref: '#/components/schemas/MeetingManagement_MeetingParticipantResponse'
        recordings:
          type: array
          items:
            $ref: '#/components/schemas/MeetingManagement_RecordingResponse'
        invitees:
          type: array
          items:
            $ref: '#/components/schemas/MeetingManagement_InviteeResponse'
    MeetingManagement_MeetingParticipantResponse:
      type: object
      properties:
        meetingId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        role:
          type: string
          enum:
            - HOST
            - PARTICIPANT
            - GUEST
        joinedAt:
          type: string
          format: date-time
        leftAt:
          type: string
          format: date-time
    MeetingManagement_CursorScrollResponseRecordingResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MeetingManagement_RecordingResponse'
        size:
          type: integer
          format: int32
        nextPageToken:
          type: string
    MeetingManagement_JsendResponseCursorScrollResponseRecordingResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: >-
            #/components/schemas/MeetingManagement_CursorScrollResponseRecordingResponse
        message:
          type: string
...